Больше не создаем ViewModel в RevitRepository
